feat: bulk2 refactor
