+++ src/connection.ts
@@ -56,7 +56,8 @@ import Process from './process';
 import { formatDate } from './util/formatter';
 import Analytics from './api/analytics';
 import Apex from './api/apex';
-import { Bulk, BulkV2 } from './api/bulk';
+import { Bulk } from './api/bulk';
+import { BulkV2 } from './api/bulk2';
 import Chatter from './api/chatter';
 import Metadata from './api/metadata';
 import SoapApi from './api/soap';
