+++ src/index.ts
@@ -2,6 +2,7 @@ import jsforce from './jsforce';
 import './api/analytics';
 import './api/apex';
 import './api/bulk';
+import './api/bulk2';
 import './api/chatter';
 import './api/metadata';
 import './api/soap';
